Add duckdb ETL pipeline
- Implemented `ingest_mimic_csv_to_duckdb.py` to load MIMIC-IV CSV files into DuckDB schemas.
- Created `load_custom_mappings_to_etl.py` for ingesting custom mapping CSVs into ETL vocabulary tables.
- Added `load_vocab_to_duckdb.py` to facilitate loading OMOP vocabulary from Athena exports into DuckDB.
- Refactored `run_workflow.py` to enhance workflow execution with better logging and variable handling.
- Developed `generate_mock_data.py` to create mock MIMIC-IV CSVs and a minimal vocabulary DuckDB for testing.
- Introduced `run_mock_pipeline.py` to automate the generation of mock data and execution of the ETL workflow.

+++ etl/etl/cdm_fact_relationship.sql
@@ -12,11 +12,11 @@
 
 CREATE OR REPLACE TABLE @etl_project.@etl_dataset.cdm_fact_relationship
 (
-    domain_concept_id_1     INT64     not null ,
-    fact_id_1               INT64     not null ,
-    domain_concept_id_2     INT64     not null ,
-    fact_id_2               INT64     not null ,
-    relationship_concept_id INT64     not null ,
+    domain_concept_id_1     INTEGER   not null ,
+    fact_id_1               INTEGER   not null ,
+    domain_concept_id_2     INTEGER   not null ,
+    fact_id_2               INTEGER   not null ,
+    relationship_concept_id INTEGER   not null ,
     -- 
     unit_id                       STRING  
 )
